requirements.txt
beautifulsoup4==4.12.2
PyPDF2==1.26.0
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
distro==1.8.0
Flask==3.0.0
greenlet==3.0.0
gunicorn==20.0.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
JPype1==1.4.1
MarkupSafe==2.1.3
numpy==1.26.0
packaging==23.2
pandas==2.1.1
pandasql==0.7.3
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.22
tabula-py==2.8.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
Werkzeug==3.0.0
